SecureSnakeYamlConstructor: Increase Scope, decrease False Positives
`new Yaml()` will now only be fixed if somewhere in scope `load*` is called on a `Yaml` instance,
or if an instance of a `yaml` object escapes the scope of the method the constructor was called.

This way, if the `Yaml` instance is just being used to dump an object to YAML, it won't be flagged/fixed.